# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
  /api/cloud/v1/orgs/{orgId}/attachments:
    post:
      operationId: prepareAttachment
      tags: [Workspace]
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareAttachmentInput"
      responses:
        "201":
          description: Attachment operation succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareAttachmentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/attachments/{attachmentId}/complete:
    post:
      operationId: completeAttachment
      tags: [Workspace]
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Attachment operation succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/attachments/{attachmentId}/read-grant:
    post:
      operationId: attachmentReadGrant
      tags: [Workspace]
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Attachment operation succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentReadGrant"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/attachments/materialize:
    post:
      operationId: materializeAttachments
      tags: [Workspace]
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterializeAttachmentsInput"
      responses:
        "200":
          description: Attachment operation succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterializedAttachments"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Reusing a key with the same command returns the original result.
        Reusing it with a different command returns an IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AttachmentIds:
      type: array
      maxItems: 8
      uniqueItems: true
      items:
        type: string
        format: uuid
    ImageAttachment:
      type: object
      required: [id, filename, size, mimeType, sha256, status]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        mimeType:
          type: string
          enum: [image/png, image/jpeg, image/webp, image/gif, image/bmp]
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        status:
          type: string
          enum: [pending, ready, expired]
    PrepareAttachmentInput:
      type: object
      additionalProperties: false
      required: [projectId, filename, size, mimeType, sha256]
      properties:
        projectId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        filename:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        mimeType:
          type: string
          enum: [image/png, image/jpeg, image/webp, image/gif, image/bmp]
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
    AttachmentUploadGrant:
      type: object
      required: [url, fields, expiresAt]
      properties:
        url:
          type: string
          description: Absolute storage URL, or development URL relative to the control-plane origin.
        fields:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          type: string
          format: date-time
    AttachmentReadGrant:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
    PrepareAttachmentResponse:
      type: object
      required: [attachment, upload]
      properties:
        attachment:
          $ref: "#/components/schemas/ImageAttachment"
        upload:
          $ref: "#/components/schemas/AttachmentUploadGrant"
    AttachmentResponse:
      type: object
      required: [attachment]
      properties:
        attachment:
          $ref: "#/components/schemas/ImageAttachment"
    MaterializeAttachmentsInput:
      type: object
      additionalProperties: false
      required: [attachmentIds]
      properties:
        attachmentIds:
          $ref: "#/components/schemas/AttachmentIds"
    MaterializedAttachments:
      type: object
      required: [paths, workerId, epoch]
      properties:
        paths:
          type: array
          items:
            type: string
        workerId:
          type: string
        epoch:
          type: integer
          format: int64